requirements.txt
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
idna==3.2
multidict==5.1.0
pyactiveresource==2.2.2
PyJWT==2.1.0
python-dotenv==0.19.0
PyYAML==5.4.1
ratelimit==2.2.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
